+++ app/portainer/registry-management/views/create/createRegistry.js
