+++ app/portainer/registry-management/views/edit/registry.html
