+++ app/portainer/registry-management/views/list/registries.html
