+++ app/portainer/registry-management/views/list/registriesController.js
